+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # v1.11.0 - ???
 
 ### Features:
-- 
+- Added !coords party command to send your current coordinates (Aliases: !c, !xyz)
+- Added !ptme alias to the !pt party command
 
 ### Changes: 
 -